Changed points from int32 to int64

+++ MegaCorpClash.Models/Company.cs
@@ -10,7 +10,7 @@ public sealed class Company
     public string ChatterName { get; set; }
     public string CompanyName { get; set; }
     public string Motto { get; set; } = "We don't need a motto";
-    public int Points { get; set; }
+    public long Points { get; set; }
     public List<Employee> Employees { get; set; } = new();
 
     [JsonIgnore]
